+++ rust/bridge/shared/testing/src/net_env.rs
@@ -81,7 +81,6 @@ pub(crate) fn localhost_test_env_with_ports(
             ports.chat_port,
             root_certificate_der,
         ),
-        chat_noise_config: None,
         chat_ws_config: RECOMMENDED_CHAT_WS_CONFIG,
         cdsi: EnclaveEndpoint {
             domain_config: localhost_test_domain_config_with_port_and_cert(