+++ rust/net/infra/src/route/describe.rs
@@ -134,14 +134,6 @@ impl std::fmt::Display for UnresolvedRouteDescription {
 }
 
 impl UnresolvedRouteDescription {
-    pub fn proxy(&self) -> Option<ConnectionProxyKind> {
-        self.proxy
-    }
-
-    pub fn domain_front(&self) -> Option<&'static str> {
-        self.front
-    }
-
     pub fn fake() -> Self {
         Self {
             front: None,